mcp/requirements.txt
fastmcp==2.14.2
python-dotenv>=1.0.0
pydantic>=2.0.0
mcp>=1.0.0
